ADD: database class which passes functions through to child classes, which implement it specifically for its sql type (currently only sqlite3), allowing more db types
FIX: html fixed/updated, uses less dv-template defaults and has distinct
classes almost everywhere now
FIX: new default style (dvbb.css), which is unfinished. Currently only
has desktop view, no mobile view

DEL: views for ajax, since we no longer load views if output is json
DEL: admin maintenance view since it's not implemented anyway and maybe
won't even be

+++ application/controller/userController.php
@@ -243,8 +243,10 @@ class userController extends controller {
                 
                 // everything else goes into settings, but first we delete existing values
                 $settings = $user->getSettings();
-                foreach ($settings as $setting) {
-                    $setting->delete();
+                if (count($settings) > 0) {
+                    foreach ($settings as $setting) {
+                        $setting->delete();
+                    }
                 }
                 // now loop through the post settings
                 foreach ($values as $key => $value) {
